requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
scikit-image
